requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
MySQL-python==1.2.5
SQLAlchemy==1.0.12
alembic==0.8.5
Jinja2==2.6
Werkzeug==0.8.3
wsgiref==0.1.2
slackclient==0.18.0
Google-Search-API==1.1.11
spacy==1.1.2